@@ -1,15 +1,7 @@
 {
   "name": "@ajf/ionic",
-  "version": "9.0.1-sha-3489afd",
+  "version": "9.0.1-sha-37fb771",
   "description": "",
-  "main": "./bundles/ionic.umd.js",
-  "fesm5": "./fesm5/ionic.js",
-  "fesm2015": "./fesm2015/ionic.js",
-  "esm5": "./esm5/index.js",
-  "esm2015": "./esm2015/index.js",
-  "typings": "./index.d.ts",
-  "module": "./fesm5/ionic.js",
-  "es2015": "./fesm2015/ionic.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/gnucoop/ajf.git"
@@ -21,14 +13,14 @@
   },
   "homepage": "https://github.com/gnucoop/ajf#readme",
   "peerDependencies": {
-    "@angular/core": "^9.0.0 || ^10.0.0",
-    "@angular/common": "^9.0.0 || ^10.0.0",
-    "@angular/router": "^9.0.0 || ^10.0.0",
+    "@angular/core": "^10.0.0 || ^11.0.0",
+    "@angular/common": "^10.0.0 || ^11.0.0",
+    "@angular/router": "^10.0.0 || ^11.0.0",
     "@gic/angular": "^5.0.0",
     "@ionic/angular": "^5.0.0",
-    "@ajf/core": "9.0.1-sha-3489afd",
+    "@ajf/core": "9.0.1-sha-37fb771",
     "flag-icon-css": "^3.3.0",
-    "tslib": "^1.9.0"
+    "tslib": "^1.10.0"
   },
   "schematics": "./schematics/collection.json",
   "ng-update": {
@@ -38,5 +30,11 @@
       "@ajf/core"
     ]
   },
-  "sideEffects": false
+  "sideEffects": false,
+  "main": "./bundles/ionic.umd.js",
+  "fesm2015": "./fesm2015/ionic.js",
+  "esm2015": "./esm2015/index.js",
+  "typings": "./index.d.ts",
+  "module": "./fesm2015/ionic.js",
+  "es2015": "./fesm2015/ionic.js"
 }
